move tests to base class
